requirements.txt
numpy
Pillow
scikit-image
pybrisque
libsvm
openai
termcolor
python-dotenv
virtualenv
opencv-python-headless